Implement parking spot detection
Car drives forward until the parking spot is found or there is an
obstacle in front of it. Algorithm calculates minimum depth for the
parking spot using L3 sensor and checks if the spot is valid based
on L5 sensor data (ignoring very bottom row, because it is noisy).
Spot is correct if there are at least MIN_CONSECUTIVE_COLUMNS in
the sensors data.

Signed-off-by: Wojciech Lewandowski <lw51032@zut.edu.pl>